上傳

+++ src/com/camp/servlet/DeleteProduct.java
@@ -2,7 +2,9 @@ package com.camp.servlet;
 
 import java.io.IOException;
 import java.sql.Connection;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -20,7 +22,16 @@ public class DeleteProduct extends HttpServlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		Map<String, String> errorMsg = new HashMap<>();
+		request.setAttribute("errorMsg", errorMsg);
 		String productId = request.getParameter("productId");
+		if (productId == null || productId.trim().length() == 0) {
+			errorMsg.put("productId", "產品編號欄必須輸入");
+		}
+		if (!errorMsg.isEmpty()) {
+			request.getRequestDispatcher("/product/GetAllProduct.jsp").forward(request, response);
+			return;
+		}
 		try {
 			ProductService pdService = new ProductService();
 			pdService.deleteOne(Integer.valueOf(productId));
@@ -31,6 +42,9 @@ public class DeleteProduct extends HttpServlet {
 			request.getRequestDispatcher("/product/GetAllProduct.jsp").forward(request, response);
 		} catch (Exception e) {
 			e.printStackTrace();
+		}if (!errorMsg.isEmpty()) {
+			request.getRequestDispatcher("/product/GetAllProduct.jsp").forward(request, response);
+			return;
 		}
 	}
